requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
WTForms==3.0.1
Flask-WTF==1.1.1
schedule==1.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.10
pymysql==1.1.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
psutil==5.9.6
requests==2.31.0